Remove ServerComponentRenderingTest.CanDispatchAsyncWorkToSyncContext
This scenario is covered by RendererSynchronizationContextTest.InvokeAsync_SyncWorkInAsyncTaskIsCompletedFirst
which is unquarantined and reliable.

Fixes https://github.com/dotnet/aspnetcore/issues/19413
